use decay rate for acid and save ParticleContextManger as class member

+++ FallingParticleSimulation/ParticleContext.cpp
@@ -25,7 +25,7 @@ Vec2 ParticleContext::getDVelocity() {
 	return dVelocity;
 }
 
-uint8_t ParticleContext::getDecayRate() {
+float ParticleContext::getDecayRate() {
 	return decayRate;
 }
